Numerics in beetween

+++ ydb/library/benchmarks/queries/tpcds/pg/q01.sql
@@ -14,7 +14,7 @@ group by sr_customer_sk
 from customer_total_return ctr1
 ,{{store}}
 ,{{customer}}
-where ctr1.ctr_total_return > (select avg(ctr_total_return)*1.2::numeric
+where ctr1.ctr_total_return > (select avg(ctr_total_return)*1.2
 from customer_total_return ctr2
 where ctr1.ctr_store_sk = ctr2.ctr_store_sk)
 and s_store_sk = ctr1.ctr_store_sk
