+++ ydb/library/benchmarks/queries/tpcds/pg/q49.sql
@@ -18,17 +18,17 @@ select  channel, item, return_ratio, return_rank, currency_rank from
  	(	select ws.ws_item_sk as item
  		,(cast(sum(coalesce(wr.wr_return_quantity,0)) as decimal(15,4))/
  		cast(sum(coalesce(ws.ws_quantity,0)) as decimal(15,4) )) as return_ratio
- 		,(cast(sum(coalesce(wr.wr_return_amt,0::numeric)) as decimal(15,4))/
- 		cast(sum(coalesce(ws.ws_net_paid,0::numeric)) as decimal(15,4) )) as currency_ratio
+ 		,(cast(sum(coalesce(wr.wr_return_amt,0)) as decimal(15,4))/
+ 		cast(sum(coalesce(ws.ws_net_paid,0)) as decimal(15,4) )) as currency_ratio
  		from 
  		 {{web_sales}} ws left outer join {{web_returns}} wr 
  			on (ws.ws_order_number = wr.wr_order_number and 
  			ws.ws_item_sk = wr.wr_item_sk)
                  ,{{date_dim}}
  		where 
- 			wr.wr_return_amt > 10000::numeric
- 			and ws.ws_net_profit > 1::numeric
-                         and ws.ws_net_paid > 0::numeric
+ 			wr.wr_return_amt > 10000
+ 			and ws.ws_net_profit > 1
+                         and ws.ws_net_paid > 0
                          and ws.ws_quantity > 0
                          and ws_sold_date_sk = d_date_sk
                          and d_year = 2000
@@ -61,17 +61,17 @@ select  channel, item, return_ratio, return_rank, currency_rank from
  		cs.cs_item_sk as item
  		,(cast(sum(coalesce(cr.cr_return_quantity,0)) as decimal(15,4))/
  		cast(sum(coalesce(cs.cs_quantity,0)) as decimal(15,4) )) as return_ratio
- 		,(cast(sum(coalesce(cr.cr_return_amount,0::numeric)) as decimal(15,4))/
- 		cast(sum(coalesce(cs.cs_net_paid,0::numeric)) as decimal(15,4) )) as currency_ratio
+ 		,(cast(sum(coalesce(cr.cr_return_amount,0)) as decimal(15,4))/
+ 		cast(sum(coalesce(cs.cs_net_paid,0)) as decimal(15,4) )) as currency_ratio
  		from 
  		{{catalog_sales}} cs left outer join {{catalog_returns}} cr
  			on (cs.cs_order_number = cr.cr_order_number and 
  			cs.cs_item_sk = cr.cr_item_sk)
                 ,{{date_dim}}
  		where 
- 			cr.cr_return_amount > 10000::numeric 
- 			and cs.cs_net_profit > 1::numeric
-                         and cs.cs_net_paid > 0::numeric
+ 			cr.cr_return_amount > 10000 
+ 			and cs.cs_net_profit > 1
+                         and cs.cs_net_paid > 0
                          and cs.cs_quantity > 0
                          and cs_sold_date_sk = d_date_sk
                          and d_year = 2000
@@ -102,15 +102,15 @@ select  channel, item, return_ratio, return_rank, currency_rank from
  	from
  	(	select sts.ss_item_sk as item
  		,(cast(sum(coalesce(sr.sr_return_quantity,0)) as decimal(15,4))/cast(sum(coalesce(sts.ss_quantity,0)) as decimal(15,4) )) as return_ratio
- 		,(cast(sum(coalesce(sr.sr_return_amt,0::numeric)) as decimal(15,4))/cast(sum(coalesce(sts.ss_net_paid,0::numeric)) as decimal(15,4) )) as currency_ratio
+ 		,(cast(sum(coalesce(sr.sr_return_amt,0)) as decimal(15,4))/cast(sum(coalesce(sts.ss_net_paid,0)) as decimal(15,4) )) as currency_ratio
  		from 
  		{{store_sales}} sts left outer join {{store_returns}} sr
  			on (sts.ss_ticket_number = sr.sr_ticket_number and sts.ss_item_sk = sr.sr_item_sk)
                 ,{{date_dim}}
  		where 
- 			sr.sr_return_amt > 10000::numeric 
- 			and sts.ss_net_profit > 1::numeric
-                         and sts.ss_net_paid > 0::numeric 
+ 			sr.sr_return_amt > 10000 
+ 			and sts.ss_net_profit > 1
+                         and sts.ss_net_paid > 0 
                          and sts.ss_quantity > 0
                          and ss_sold_date_sk = d_date_sk
                          and d_year = 2000
